Fixing deprecated usages of old numpy dtypes.
Removed references to:
- "numpy.int" and "numpy.long" (replaced by "int"),
- "numpy.float" (replaced by "float")
- "numpy.bool" (replaced by "bool")

References to specific type sizes (e.g. "numpy.float32" or "numpy.int16") are not deprecated, and remain in place.
Removed the requirement in environmental.yaml to suggest older versions of numpy, for which this deprecated code didn't break things. This code should now be compatible with the latest versions of numpy (as of this commit).